Improve webrequest-refine query
The patch enforces the DISTINCT part of the query to be made
on raw data only, instead of raw data plus augmented values
computed on raw data. This reduces significantly the data to be
shuffled (by half on my test), reducing both network and disk IOs.
It also materializes reused fields in the reducer stage, preventing
recomputation at every reuse.
Finally it grows the number of buckets of the table from 64 to 256
to increase parallelization of the reduce step as most of the
computation will be made during that stage from this patch. A side
effect of this is to grow the number of files for every webrequest
hour - this impacts the namenode as more files costs more than more
blocks for less files.

Bug: T267008
Change-Id: I45401ee947a1dd0dc80c40c3548fc3454dbcf33b