Add a parameter to refinery-deploy-to-hdfs to force umask to 022
The current HDFS umask settings are 027, so other doesn't have any
read permissions on new files created. When we upload refinery to HDFS
we need to ensure that several users, like oozie, can read the files.
To simplify the handling of permissions, this patch adds a custom
umask setting (022) to hdfs dfs commands in order to override the
default settings and avoid issue during deploy and/or oozie job restarts.

Bug: T270629
Change-Id: I5d9de4e8f3c727abfc05b3cf232c868546c25ae8