Update permissions of oozie jobs writing archives
We have changed the HDFS default umask to 027, meaning that by defaut
OTHER don't have any access to data. For archived data, we need OTHER
to have access by default, as archived data is mostly used for data to
be made available externally.
This patch makes the archive subworflow change permissions to readable
by OTHER by default, add explicit configuration for jobs to which those
defaults don't fit, and makes use of a HDFS config change for the
mediawiki-history-dumps to achive the same, as this job doesn't use the
archive subworkflow.

Bug: T270629
Change-Id: If3ad23c6ed36b86871a6de0f8d7fc2847826e84e