HivePartition.list must return valid HQL
The analytics cluster has recently upgraded to hive 2.x which has a
new set of reserved keywords, including those used in partitioning
such as `date` and `timestamp`. To ensure the process can continue to
work all column names must be quoted with backticks so Hive
understands the intent that these are columns and not keywords.

Change-Id: I52f15889915bd7f372459f58786607e8f16996de