bin/camus - use hadoop classpath when running checker jar.
Because refinery-camus.jar is not an 'uber' jar (like refinery-job), it does
not have Hadoop dependencies built in.  Previously we just globbed in /usr/lib/spark/jars/*
and pulled in Hadoop deps.  This morning, hadoop jars were removed from that directory.

This uses hadoop classpath to put the proper jars on the classpath.

Bug: T274384
Change-Id: Ie5a21465e528b3404a9deb5010d1ce08e1ada716