Fix and optimize Hive query and change field names in properties file for top-per-country job
JSON output by the top-per-country Oozie job Hive query was previously malformed due to a missing double quote. This change corrects that issue.
This change also modifies the query to be more efficient and cleaner.
Additionally, this change modifies the naming scheme of the Cassandra fields to reduce and simplify processing on the AQS side.

Bug: T207171
Bug: T263697
Change-Id: I4816d38258b5fd614a928621ff42d0a4112241f5