Don't use Gobblin lock but rather yarn check
Gobblin has a built in lock mechanism that allows for a job not to start
if a job of the same name is already running. The problem of this approach
is that since we don't (yet) have gobblin metrics, we don't know if a job
is stuck or not, as the job not starting doesn't fail (return code 0).
This patch disables using Gobblin lock and reimplements the original
check in the python wrapper, calling Yarn and checking if a job with the
same name already runs and actually failing if so.

Bug: T286559
Change-Id: Idabcaf09ab454bbdbd9f2bdb999a616e309c7787