Grow mediawiki-history oozie jobs resources
For mediawiki-history-denormalize job we provide twice as much ram (2.4Tb)
over more than two times the number of nodes (80 instead of 32).
This speeds up the execution and facilitates shuffled data to be spread
among workers.

For checkers and dumps jobs we grow the number of executors (32 to 80) but reduce
RAM and executor-cores to keep resource footprint small.
For checkers we also added a new num-partitions parameter to fluidify computation
through smaller tasks. This addition means updating the underlying job jar to a
new version, hence using its new '-shaded' version.

Bug: T290469
Change-Id: I34a668658f727445d3b2d0b474b57dccd17e9a68