gobblin: use the new jks TLS bundle to validate certificates
On all hosts we now deploy
the /etc/ssl/localcerts/wmf_trusted_root_cas.pem bundle,
that is composed by the Puppet CA and the Root PKI CA certs.
This bundle is useful to accept TLS certificates signed by
both CAs (or their Intermediates, like the new Kafka one).
It should guarantee a smooth and uneventful experience to
Kafka clients while we'll swap TLS certificates.

Gobblin seems to require a JKS bundle as truststore,
so this change needs a puppet change that enables the
deployment of the jks truststore on the nodes where
Gobblin runs (like an-launcher1002).

Bug: T291905
Change-Id: Icb895d5c483654ffe21c1edbcf648204c2897a09