Create discussiontools_subscription table in Hive
This is the script that creates the
mediawiki_private_discussiontools_subscription table in Hive.
The table will receive sqoop output on a daily basis.  It's a little bit
different from all the other sqoop jobs in that it's always writing to
the "latest" partition.  Once the table is creation (on deployment of
this change), and one round of data is sqooped, and msck repair table
called, updates should just work from then on.  This is because they
overwrite data on the existing partitions, almost never creating new
partitions.

NOTE: This approach will get out of sync with the data when new wikis
are added, so a manual msck repair will have to be issued at that point,
and probably from time to time.  Since this is just an experiment for
now, we're willing to live with the technical debt.

Bug: T290516
Change-Id: I4574b1df0294ec486f64f6f2c5b07d1add823347