gobblin: move to the new canonical bundle jks location
The file contain the same certs as the previous one.

For the moment this code is fine, but if we want to make
it usable/transparent for deployment-prep and pontoon,
it is better that this configuration file gets rendered
by puppet instead (in order to use the variables exposed
by profile::base::certificates).

Bug: T296089
Change-Id: I0c93294f666b06f98daba4f102804f83f21ce286