Remove mediawiki_skin_diff from allow list
Initially this schema had a problem because `keep` is not an allowed
sanitization mechanism for structs, and it was applied to
`normalized_host`.  I changed it to `keep_all`, but there seems to be a
bug in SanitizeTransformation.scala around line 400 that prevents that
from being used also.  For now, disabling the schema until the refine
fix goes through.

Change-Id: I468f7d9781a2739d5566f023d13b5c9c14c3cfc0