Add 2010 wikitext editor option to edit_hourly interface field
A new revision tag ('wikieditor') was created to tag edits made with the 2010 wikitext editor
in T249038. We want to add this as an option to the current interface field in
the edit_hourly druid dataset in Superset so that users can view and filter edits
made with the 2010 wikitext editor.

Bug: T293406
Change-Id: I8cf180d50faf7a20286360b094a20da880412a22