Throttle heavy monthly jobs
4 monthly jobs (mediarequest and pageview top for cassandra, 2 unique-devices)
need to shuffle a lot of data. This shuffling puts pressure on the network up
to link saturation, impacting production traffic until we have QoS setup.
This patch throttles the number of reducers the jobs can run in parallel
to prevent them fetching too much data at once.
Note: The patch doesn't lower the number of reducers the jobs use, it lowers
the number of reducers that can be run at the same time. This leads to a longer
job with less network pressure, but with the level of parallelisation.

Bug: T307779
Change-Id: I78d4111577b32de360966bdbb826fd53a67376b8