Track accesses to Special:EntitySchemaText/* on Wikidata
`uri_host = 'www.wikidata.org'` was used over
`normalized_host.project_family = 'wikidata'` because it seems to be
about twice as fast in practice.

`uri_path like '/wiki/Special:EntitySchemaText/%'` was used over
`x_analytics_map['special'] = 'EntitySchemaText'` because apparently the
`special` field does not actually exist in that map, at least not for
that special page. It appears that field is only set if the page opened
is an actual Wiki page, not just downloading a plain text file.

The GROUP BY is used on the date columns directly as there was some
weirdness around using them with the `ts` alias.

Bug: T304793
Change-Id: I9090ddea1c9392ee80bd58f0499d9f9a7e20ac68