Make wikidata reliability metrics longs so they fit in graphite
When we migrated the wikidata reliability metrics from Spark to
SparkSql we forgot to transform the time_firstbyte metric from
a double to a bigint (and change it from sec to ms). Because
of that, graphite would not accept it. See line that we missed:

https://github.com/wikimedia/analytics-refinery-source/blob/master/refinery-job/src/main/scala/org/wikimedia/analytics/refinery/job/wikidata/WikidataReliabilityMetrics.scala#L144

This change fixes this by making the metric a bigint, and multi-
plying it by 1000 (sec to ms).

Bug: T314131
Change-Id: I2ce59e21e85014f13b064064a11b682f34eaacab