Get actor information from actor table
The checkuser table (cu_changes) has been updated to always include
cuc_actor, a foreign key to the actor table.  The easiest way to adapt
our sqoop code is to just join and get the information from the actor
table directly.  If this does not perform well, we'll have to get the
actor_id and join later, which means some more significant oozie job
changes.

Bug: T326330
Change-Id: If239ff338d93e2c84fec7951346681afd92f8f13