Add unit tests on raw webrequest data loss reports job
The bin script `refinery-drop-older-than` will be used for one
more case: removing data loss reports files on HDFS. Those files are
generated for each hourly partition of webrequest.

In this changeset, unit tests are added to make sure the new call will
do what it is supposed to do for the new use case.

Bug: T332707
Change-Id: I4a6adb723e3f37b7f253b1eeab54253f69a51ee1