Migrate banner activity druid loading to Spark3 and Airflow
The first patch set only copies the old query to the new location,
and the second patch set applies the changes, so it's easier to review,
and also adds the monthly query (previously just recompaction).
Note that Spark3 doesn't seem to support CREATE MACRO, so I had to
remove the MACRO and play a bit with CTEs instead.

I tested the queries manually, and they perform well. The generated
data matches the one generated by the old query.

Bug: T336184
Change-Id: Ic17d1d21992de21f7bc76af0c97958497267985b