Create wikilambda_zobject_labels & wikilambda_zobject_function_join tables in Hive
[Comment copied from
https://gerrit.wikimedia.org/r/c/analytics/refinery/+/739922:]

This script creates the wikilambda_zobject_labels and
wikilambda_zobject_function_join tables in Hive.
The tables will receive sqoop output on a regular basis.  These are a bit
different from other sqoop jobs in that they are always writing to
the "latest" partition.  Once the table is created (on deployment of
this change), and one round of data is sqooped, and msck repair table
called, updates should just work from then on.  This is because they
overwrite data on the existing partitions, almost never creating new
partitions.

NOTE: This approach will get out of sync with the data when new wikis
are added, so a manual msck repair will have to be issued at that point,
and probably from time to time.  Since this is just an experiment for
now, we're willing to live with the technical debt.

Bug: T341728
Bug: T341729
Change-Id: Id5da467afcacb0f82da26165e2d98dbb7b3f7b62