Add client_dt to EditAttemptStep allowlist
We would like to add client_dt to EAS allowlist, so that it can be used for analyses on historic data. The dt field which is currently allowed in the sanitized version is logged server-side, it would be useful to have client_dt in the sanitized version as well, especially when analysing events within a session.

Bug: T341888
Change-Id: I7924eff152e6b576bb3b520a9b347a4f55fb6784