Use sudo with git in refinery_deploy_to_hdfs
This change modifies the refinery_deploy_to_hdfs script such that it
calls sudo and runs 'git describe' as the analytics-deploy user, rather
than the user who is running the script.

The reason for this is to address a problem caused by the fact that the
git repository is owned by analytics-deploy and recent versions of git
complain about this discrepancy.

Suitable sudo rules have been added to the hosts in puppet here:
https://gerrit.wikimedia.org/r/c/operations/puppet/+/950194

Bug: T334493
Change-Id: I45530b09d3943de769b3ccf51610bb4d1676ca33