Update refinery-deploy-to-hdfs to use sudo
We found that the recent changes to the refinery-deploy-to-hdfs script
didn't work because we were already using sudo outside the script in
order to impersonate the hdfs user.

This change updates the script so that the use of sudo to become hdfs is
performed within the script instead. This allows us to impersonate the
analytics-deploy user when running git commands and the hdfs user when
running hdfs commands.

Bug: T334493
Change-Id: I96622a0b0123db6361d9ef39e5692f686bdeeeba