Add unique-devices Iceberg schemas and scripts
The first step to migrate the unique-devices tables to Iceberg is to
duplicate the existing tables and loading jobs onto iceberg tables.
This patch contains the iceberg table creation scripts, HQL queries
to regularly update them fron old hive tables,and backfill queries
to bottstrap them.
Later the old hive scripts will be deleted, and the iceberg scripts
updated to embed computation, as well as renamed.

Bug: T347689
Change-Id: I17b6585f94f435858f55cf6c1561fdc7946af467