Use canonical_data.countries when populating the referer tables
instead of the previously hardcoded list.

The countries table is joined to the request with a broadcast
join to minimize performance impact.

More protected countries are defined in the countries
table than in the hardcoded list. We anticipate
a diminution in the size of the resulting table (~15%).

A country not present in the canonical table is supposed to be
unprotected.

I've checked the modifications by running the Hive table script on
the cluster.

Bug: T348504
Change-Id: If64dcef214a24d6ea728f5df8ddc36a5f8b04e86