Quick fix to refine_webrequest_hourly for exclude_row_ids
The old code was parsing the list of row ids (hostname, sequence)
into a map. But when there were several corruped rows with the
same hostname, that raised a duplicated map key error.

This change changes the expected format of the list of row ids,
and parses them as an array, and later splits the strings into
hostname and sequence.

Change-Id: I0d6ed7c3007c261895dcd4bf5362f33e01db1005