Unique devices druid ingestion job - Iceberg migration
Migrate the source of unique devices Druid ingestion job to Iceberg the
newly created Iceberg tables from the hive tables. It's performed with
an update of the HQL files.

I've also included a cleanup of the `airflow/` dir which was a remain
of a previous poc.

There is a companion Gitlab patch in airflow-dags to deploy jointly.

Bug: T347879
Change-Id: Ibe0f7bde172e18c7225a5aa9e0467655081e6203