Adopt a more resilient approach to use webrequest x-analytics
The WebRequest x-analytics field is a dictionary serialized as a String.
Pairs are separated with semi-colons.
Keys and values are separated with equal signs.

The historical approach to retrieve data within this String was to use a regex.

We are now using a more resilient approach based on the Spark SQL function str_to_map.

A companion patch has been created in refinery to make use of this patch:
https://gerrit.wikimedia.org/r/c/analytics/refinery/source/+/992475

Bug: T355391
Change-Id: I90ebb98c4acf42d4c41fae54ab48af5a9725bd9c