Add app_donor_experience to allowlist
To preserve analytics data related to Fundraising efforts on the
mobile apps (iOS and Android). All the identifiers are hashed
per sanitization guidelines.

This allowlist entry includes geocoded data down to country level
as we need to compare across countries. This is similar to the
Fundraising team's need to keep country-level geo data for the
WikipediaPortal schema.

Bug: T356214
Change-Id: I4fc755e112d1c626a74ea0b08850c84d329f44d1