Switch to using the new redacted replica MariaDB server
We have previously been using clouddb1021 as a private copy of the
clouddb replicas. During a refresh we have taken the opportunity to
rename this, for clarity, to an-redacteddb1001.

This commit causes the new server to be used by the refinery.util
package and the analytics-mysql utility script.

We have renamed the corresponding variables and arguments, replacing
reference to 'cloud' with 'redacted' instead.

Bug: T365453
Change-Id: Ie3de42c041a9c5138e9c7343bd4d6f440840bea2