Implement new way to aggregate browser statistics
NOTE: use "Redacted" as the value for the unknown parameters when
updating the corresponding Airflow job.

NOTE: we looked into what we could possibly do to pre-aggregate
pageview_hourly to facilitate faster versions of this and other queries.
The problem is that most of the data comes from the long tail of unique
rows in pageview_hourly.  So any pre-aggregation did not seem easy to do
and still serve multiple purposes.

Bug: T342267
Change-Id: If83f0fd68b61d013b1ccd73c65590f799cc7b85f