Modify the automated traffic detection pipeline to include redirects
This change makes the webrequest_actor pipeline aware of redirects to pageviews.
This way, we will not only classify the actor signatures by the metrics over
their pageview requests, but also by the metrics over their redirect requests.

If an actor is classified as automated either via pageviews or via redirects
(or both), it is considered an automated actor.
That said, the pipeline keeps metrics about pageviews separate from redirects,
to not affect the tuning of the classifier heuristics.

Bug: T375527
Change-Id: I2c5eec0a666b8b9ee8c130ce1cc12122410ca168