Modify MediaWiki History queries to support Temp Accounts
This change adjusts the MediaWiki History Reduced query to support Temp Accounts.
Actually, it makes it so that temporary users are labeled as anonymous.
This was decided to make the query and all the downstream pipelines backwards
compatible.

This change also updates the create table statements affected by the changes
in the MediaWiki History scala code.

Bug: T379230
Change-Id: I8b108f283c4002876cb71fc9cbbef2e3ee3368ff