Optimize Commons Impact Metrics calculations by using digests
After a sudden growth in some Commons categories, the Commons
Impact Metrics calculations for commons_category_metrics_snapshot
have started failing.

The error is caused by a too large collect_list operation, which
tries to gather the page titles of 32M pages in a single list.

This optimization makes use of md5 digests to encode the mentioned
page titles in shorter strings while keeping their identities.
This is possible, because those page titles were solely used to
calculate distinct counts.

The code substrings the mdf digest to the first 10 characters,
which has the potential to encode 1 trillion of page titles.

This change has been tested (with the offending category)
and it worked.

Bug: T381799
Change-Id: Ia4c618351f5bb802142f5715025b0167f862b6d2