Add columns to data_quality_alerts to support inserting ResultKey
We need a more specific way of identifying where a verification check belongs to, since these checks can be computed against different subsets of a source_table.
This patch lets us insert the same ResultsKey that a metrics computation might use so we can match them better.

This will require us to ALTER the alerts table before it can be used.

Bug: T384962
Bug: T389162
Change-Id: If491f793876f7e7b4836b59d098665f1165ce66c