Update Gobblin webrequest_frontend timestamp
The Gobblin webrequest_frontend job was using the Kafka Timestamp
to partition the data. This timestamp is generated at the end of the
request when HAProxyKafka sends the message.
This generates a discrepency between sequence number (generated at
request start) and the timestamp used for bucketing data, leading to
the sequence-validation algorithm failing.

Bug: T388397
Change-Id: I87d7daf05c5e22aa0f3b985b97d989b0a4bfdffc