Update webrequest_frontend validation
HAProxyKafka has a different special that VarnishKafka had.
Instead of logging rows with no timestamp in case of a bad request,
it sets its sequence number to 0 and certain fields to '<BADREQ>' or
nul value.
This patch updates our validation algorithm to cover for this case
instead of the VarnishKafka on.

Bug: T389797
Change-Id: I38e673c3278f8eddce3f0e6a8756b0e66e23e482