Fix commons impact metrics all-wikis cassandra loading queries
We discovered that there was a bug in 2 cassandra loading queries
for Commons Impact Metrics: top_viewed_categories_monthly and
top_viewed_media_files_monthly. They were missing a group by
in the section compiling data for the all-wikis option.
Because of that, the query would generate duplicate rows,
that would be lost at Cassandra loading time.

This change adds the missing group by's in both query files.

Bug: T370470
Change-Id: I31c9920677fd59fab2c7b5fbe5a7ef8854c1f57d