Improve the automated traffic detection pipeline
This change adds another automated traffic detection approach
to the existing pipeline. It uses diff IP groups[1] to identify
sets of IPs that behave like crawlers/spiders.

[1] Diff IP group: The set of distinct IPs of the actors that requested
diff pages (as in editing history diff) for a particular
page_title, during a particular hour, and from a particular country.

This change adds 2 heuristics to the current algorithm:
One is the number of diff IP groups every actor belongs to,
and the other is the maximum size of the actors diff IP groups.

If an actor belong to too many diff IP groups or has participated
in a too big diff IP group, it is considered automated traffic.

Change-Id: If05a4e78af554877fd064deb770dc9f11d4165bd