Add access_method to unique-devices per-domain
The removal of the `.m` subdomain forces us to adapt our unique-devices
per-domain handling. We add the access_method field to the table to
replace the missing domain, and will backfill using `.m` subdomain.
Note: As we change unique-devices jobs, we take advantage of this patch to
remove to the old non iceberg hive tables.

Bug: T401666
Change-Id: Ibdb725502f5ac3102c37f1ea863af45bb6cb97bc